emerge: add --changed-deps-report option (bug 645780)
The --dynamic-deps=n default causes confusion for users that are
accustomed to dynamic deps, therefore add a --changed-deps-report
option that is enabled by default (if --usepkgonly is not enabled).

The --quiet option will suppress the report if none of the packages
having changed dependencies are in the dependency graph, since they
are harmless in that case. If any of these packages *are* in the
dependency graph, then --quiet suppresses the big NOTE section of
the report, but the HINT section is still displayed since it might
help users resolve problems that are solved by --changed-deps.

Example output is as follows:

!!! Detected ebuild dependency change(s) without revision bump:

    net-misc/openssh-7.5_p1-r3::gentoo
    sys-fs/udisks-2.7.5::gentoo

NOTE: For the ebuild(s) listed above, a new revision may be warranted if there
      has been a dependency change with significant consequences. Refer to the
      following page of the Gentoo Development Guide for examples of
      circumstances that may qualify:

          https://devmanual.gentoo.org/general-concepts/ebuild-revisions/

      If circumstances qualify, please report a bug which specifies the current
      version of the ebuild listed above. Changes to ebuilds from the 'gentoo'
      repository (ending with '::gentoo') can be browsed in GitWeb:

          https://gitweb.gentoo.org/repo/gentoo.git/

      Use Gentoo's Bugzilla to report bugs only for the 'gentoo' repository:

          https://bugs.gentoo.org/

      In order to suppress reports about dependency changes, add
      --changed-deps-report=n to the EMERGE_DEFAULT_OPTS variable in
      '/etc/portage/make.conf'.

HINT: In order to avoid problems involving changed dependencies, use the
      --changed-deps option to automatically trigger rebuilds when changed
      dependencies are detected. Refer to the emerge man page for more
      information about this option.

Bug: https://bugs.gentoo.org/645780