TRACK_KEYWORDS implementation.
The TRACK_KEYWORDS global allows one arch to track another arch. This
way, say, arm64 can track amd64 and have similar keyword masking.
TRACK_KEYWORDS of "amd64 ~amd64" on an unstable arm64 system will mean
that we effectively have an ACCEPT_KEYWORDS of "amd64 ~amd64 arm64
~arm64".

What is useful about TRACK_KEYWORDS as opposed to simply using
ACCEPT_KEYWORDS for this purpose is that it's still possible to
explicitly KEYWORD mask ebuilds for the native arch using -ARCH.  So it
is still possible to tag ebuilds that are known-incompatible with the
native arch via ie. -arm64. This works in profiles as well as in ebuilds
themselves.

This presents an interesting alternative to using independent arch teams
for maintaining ebuild KEYWORDS, when most of the KEYWORDS can simply
track another arch.