find_smallest_cycle: fix ignore_priority (bug 756199)
We can't use ignore_priority in gather_deps, since it will drop children
that we need to select. For find_smallest_cycle we need stop passing the
ignore_priority parameter to gather_deps, and gradually increase the
ignore_priority used to select the mergeable_nodes set.

TODO: rebase this patch on master and clean it up

Bug: https://bugs.gentoo.org/756199
Signed-off-by: Zac Medico <zmedico@gentoo.org>