_pkg_str: add binpkg_format property
Returns the BINPKG_FORMAT metadata. A return value of None means
that the format is unset. If there is no metadata available or the
BINPKG_FORMAT key is missing from the metadata, then raise
AttributeError.

In bintree.py, update code to use the _pkg_str.binpkg_format
property where appropriate. This eliminates many redundant calls
to the get_binpkg_format function.

In the binarytree.getname method, slightly refactor logic which
should behave the same as before. It should raise an error if the
desired binpkg_format is not clear. The caller should either set
allocate_new to True or else ensure that cpv.binpkg_format is set
to a particular format.

Signed-off-by: Zac Medico <zmedico@gentoo.org>