find_smallest_cycle: finer grained buildtime prioritization
When installed packages must be used to break buildtime cycles,
increase the priority as differences between installed and
new packages increase. The following differences are considered,
in order of increasing priority:

* New ebuild revision
* New package version
* New ebuild slot/subslot
* New ebuild slot/subslot pulled in by := slot operator

The included unit test case demonstrates that a direct circular
buildtime := slot operator dependency is given higher priority
when a new slot/subslot has been pulled in for the matched
package. The same test case fails for the master branch, meaning
that it would rebuild a package against an older slot/subslot,
even though it should give priority to rebuild it against the
new slot/subslot.

Bug: https://bugs.gentoo.org/754777
Signed-off-by: Zac Medico <zmedico@gentoo.org>