output: add 'g' symbol to indicate remote binary with --getbinpkg
Provide visual feedback to the user as to whether a selected binary
package is local or remote. The symbol 'g' is used to indicate that
a remote binary will be fetched due to --getbinpkg, and is shown in
the same place as 'f' or 'F'. These flags represent a similar concept
yet should never shown for remote binaries, but will still be shown
preferentially should they somehow be applicable.

Signed-off-by: Jethro Donaldson <devel@jro.nz>
Part-of: https://github.com/gentoo/portage/pull/1527
Signed-off-by: Sam James <sam@gentoo.org>